Revert "use memo"
This reverts commit 26e4000b68035480f8af79ab5114bb3ba6a0fa91.
